requirements.txt
langchain
langchain-openai
langchain-community
openai
sqlalchemy
streamlit
fastapi
uvicorn
python-dotenv
pandas
tiktoken